+++ e2e/tests/extension/a11y.spec.ts
@@ -1,0 +1,16 @@
+import { test } from '@playwright/test';
+import { config } from '../utils/config';
+import { expectA11yCompliant, login } from '../utils/helper';
+
+test('accessibility', async ({ page }) => {
+  await test.step('login page accessible', async () => {
+    await page.goto(`${config.URL}/login`);
+    await page.getByRole('button', { name: 'Login' }).waitFor();
+    await expectA11yCompliant(page);
+  });
+
+  await test.step('main page accessible', async () => {
+    await login(page);
+    await expectA11yCompliant(page);
+  });
+});