+++ services/reis/rei_s/config.py
@@ -68,8 +68,10 @@ def check_valid_s3_bucket_name(bucket_name: str | None) -> None:
         raise ValueError("FILE_STORE_S3_BUCKET_NAME cannot be formatted like an IP address")
 
 
-class Config(BaseSettings, frozen=True):
-    model_config = SettingsConfigDict(env_file=".env", env_file_encoding="utf-8", extra="ignore", env_ignore_empty=True)
+class Config(BaseSettings):
+    model_config = SettingsConfigDict(
+        frozen=True, env_file=".env", env_file_encoding="utf-8", extra="ignore", env_ignore_empty=True
+    )
 
     workers: Annotated[int, Field(gt=0)] = 1
     metrics_port: Annotated[int, Field(ge=0)] = 9200